refactor(test): 테스트 의도를 고려하여, 테스트 코드 재 작성
- 장기판 기본 이동 규칙에 대한 테스트 코드는 JanggiBoard에서 진행하되, @Nested를 활용
